no kill in windows

+++ endless.go
@@ -227,7 +227,9 @@ func (srv *endlessServer) ListenAndServe() (err error) {
 	srv.EndlessListener = newEndlessListener(l, srv)
 
 	if srv.isChild {
-		syscall.Kill(syscall.Getppid(), syscall.SIGTERM)
+		if runtime.GOOS != "windows" {
+			syscall.Kill(syscall.Getppid(), syscall.SIGTERM)
+		}
 	}
 
 	srv.BeforeBegin(srv.Addr)
@@ -278,7 +280,9 @@ func (srv *endlessServer) ListenAndServeTLS(certFile, keyFile string) (err error
 	srv.EndlessListener = tls.NewListener(srv.tlsInnerListener, config)
 
 	if srv.isChild {
-		syscall.Kill(syscall.Getppid(), syscall.SIGTERM)
+		if runtime.GOOS != "windows" {
+			syscall.Kill(syscall.Getppid(), syscall.SIGTERM)
+		}
 	}
 
 	srv.BeforeBegin(srv.Addr)
